Database Files
Can import the Dump into MySQL Workbench if need be. Queries are included in the dump but I have also included them in their own separate files which you can open with notepad if you want to see how they work. Can also be imported to MySQL but not really necessary as they are included in the dump.